role resolver updated

+++ libs/auth/src/modules/role/role.resolver.ts
@@ -1,5 +1,7 @@
+import { Action } from '@libs/auth/constants';
+import { CaslAbility } from '@libs/auth/decorators/current-user';
 import { JwtAuthGuard } from '@libs/auth/guards/jwt-auth-guard';
-import { UseGuards } from '@nestjs/common';
+import { UnauthorizedException, UseGuards } from '@nestjs/common';
 import { Args, Int, Mutation, Query, Resolver } from '@nestjs/graphql';
 import { CreateRoleInput, UpdateRoleInput } from './dto/role.dto';
 import Role from './role.entity';
@@ -11,22 +13,47 @@ export class RoleResolver {
   constructor(private readonly roleService: RoleService) {}
 
   @Query(() => [Role])
-  async roles(): Promise<Role[]> {
-    return this.roleService.findAll();
+  async roles(@CaslAbility() ability): Promise<Role[]> {
+    if (ability.can(Action.Read, 'role')) {
+      return this.roleService.findAll();
+    } else {
+      throw new UnauthorizedException();
+    }
   }
 
   @Mutation(() => Role)
-  async createRole(@Args('data') role: CreateRoleInput): Promise<Role> {
-    return this.roleService.create(role);
+  async createRole(
+    @Args('data') role: CreateRoleInput,
+    @CaslAbility() ability,
+  ): Promise<Role> {
+    if (ability.can(Action.Create, 'role')) {
+      return this.roleService.create(role);
+    } else {
+      throw new UnauthorizedException();
+    }
   }
 
   @Mutation(() => Role)
-  async updateRole(@Args('data') role: UpdateRoleInput): Promise<Role> {
-    return this.roleService.update(role);
+  async updateRole(
+    @Args('data') role: UpdateRoleInput,
+    @CaslAbility() ability,
+  ): Promise<Role> {
+    if (ability.can(Action.Update, 'role')) {
+      return this.roleService.update(role);
+    } else {
+      throw new UnauthorizedException();
+    }
   }
 
   @Mutation(() => Int)
-  async deleteRole(@Args('id') id: number): Promise<number> {
-    return this.roleService.delete(id);
+  async deleteRole(
+    @Args('id') id: number,
+    @CaslAbility() ability,
+  ): Promise<number> {
+    if (ability.can(Action.Delete, 'role')) {
+      return this.roleService.delete(id);
+    } else {
+      throw new UnauthorizedException();
+    }
   }
 }
